[TAN-867] Use USERID key for employees. Add shared specs
